11/27 update

+++ GA-CarArrangementSystem-API/DTO/RouteInfoDTO.cs
@@ -11,7 +11,7 @@ namespace GA_CarArrangementSystem_API.DTO
         public string RouteName { get; set; }
         public string RouteStart { get; set; }
         public string RouteEnd { get; set; }
-        public int RouteCostTime { get; set; }
+        public double RouteCostTime { get; set; }
         public Boolean RouteOneWay { get; set; }
         public Boolean RouteRoundTrip { get; set; }
         public string RouteRemark { get; set; }
